Dice command logic restructured, subreddit finder output modified.
Using RegEx to parse the dice input now seems to make for cleaner parsing. Allows us to fetch the exact rolls we want for the operation we're doing without having to worry about an odd order of input.

Subreddit finder output just returns a link or 'not found' message without a header.